fix api spec and apply more review suggestions
