--- airflow/api_connexion/schemas/view_menu_schema.py
@@ -1,48 +1,0 @@
-# Licensed to the Apache Software Foundation (ASF) under one
-# or more contributor license agreements.  See the NOTICE file
-# distributed with this work for additional information
-# regarding copyright ownership.  The ASF licenses this file
-# to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance
-# with the License.  You may obtain a copy of the License at
-#
-#   http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing,
-# software distributed under the License is distributed on an
-# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-# KIND, either express or implied.  See the License for the
-# specific language governing permissions and limitations
-# under the License.
-
-from typing import List, NamedTuple
-
-from flask_appbuilder.security.sqla.models import ViewMenu
-from marshmallow import Schema, fields
-from marshmallow_sqlalchemy import SQLAlchemySchema, auto_field
-
-
-class ViewMenuCollectionItemSchema(SQLAlchemySchema):
-    """ViewMenu Schema"""
-
-    class Meta:
-        """Meta"""
-
-        model = ViewMenu
-
-    id = auto_field()
-    name = auto_field()
-
-
-class ViewMenuCollection(NamedTuple):
-    """ViewMenu list"""
-
-    view_menus: List[ViewMenu]
-    total_entries: int
-
-
-class ViewMenuCollectionSchema(Schema):
-    """ViewMenu list schema"""
-
-    view_menus = fields.List(fields.Nested(ViewMenuCollectionItemSchema))
-    total_entries = fields.Int()
